Factor out buffer handling and port DHCP.
Factor out parsing from a buffer into a separate file, so I can use it
for radvd.

DHCP used an iterator for parsing, which turned out to be a bit messy,
so I've swapped out to just using a slice and an offset.

This version should hopefully do fewer bounds checks.