Add more debugging, and a test.
Occasionally erbium "forgets" about addresses on an interface.  Existing
logging shows that this occurs when it receives a NewLink message for an
existing interface.  Previously this would overwrite the interface, and
lose the associated addresses, but as a previous attempt to fix this bug
I had added support to copy the addresses from the old interface to the
new interface.  However, this bug still seems to present itself.

This is an attempt to fix that:
 - Clarify the logging, so more information can be gathered next time
   this occurs.
 - Add a test to cover this case (the test passes, despite evidence to
   the contrary in the wild).