Add tests for config file examples.
Parse out the examples from the config man page, then run them through
the config parser to make sure that they at least parse correctly.  It's
embarrassing to have mistakes in the example configs!

Also fix the missing "apply-" in the example configuration for
dns-servers in the DHCP configuration.  (Fixes: #11)