Report cli usage information in more cases.
This now outputs usage information if there is more than one argument,
or if the first argument starts with a "-".

This also returns non-exit-0, so that if we add command line arguments
in the future, and somehow someone gets rolled back to an old version of
the binary, then the program will exit unsuccessfully, which hopefully
will make it clearer what happened (esp if someone has some kind of
integration test).

This continues to improve the situation in issue #2, but I don't think
it really resolves it yet.  I've not yet come to a conclusion about what
I want to do with command line arguments.