Bump nix version to 0.25
This is a massive refactor, as the nix sockaddr types have all changed,
so I've introduced some address aliases for erbium to use, so we can be
consistent.

This involved creating a bunch of helpers.  I suspect we've also lost
a bunch of functionality along the way around unix domain sockets.

I intend to tidy this up a bit more, but this is the minimal change to
commit to get things compiling again.