Replace rust-crypto with Hmac + Sha2
rust-crypto pulls in a massive pile of dependencies, and has frequent
dependabot alerts, despite the fact we only use Hmac<Sha2>, so just
use that instead. (Also it's already in debian)

hmac seems to have a bunch of weird types to deal with the fact that
const generics weren't stable in rust, I'm looking forward to that
going away so I can simplify some stuff.